==> codes/04_Coordinates_Acquisition_with_herepy.ipynb <==
# 04. Coordinates Acquisition with herepy   

The HERE Geocoder API is a REST (Representational state transfer) API that allows us to obtain addresses from locations' coordinates or coordinates from addresses or landmarks. We are using herepy which is a library that allows us to run HERE Geocoder API in Python. 
In this project, we submit an free form address to request the corresponding coordinates.

## Table of Contents
- [Installation](#Installation)
- [Loading Libraries & Data](#Loading-Libraries-&-Data)  
- [Coordinates Acquisition](#Coordinates-Acquisition)
- [Source](#Source) 

## Installation  



We can install herepy (library that uses HERE API in python environment) using:
```Terminal
$ pip install herepy
```
Getting the code at https://github.com/abdullahselek/HerePy   
```Terminal
$ git clone git://github.com/abdullahselek/HerePy.git
$ cd HerePy

# To install dependencies, run either:
$ pip install -Ur requirements.testing.txt
$ pip install -Ur requirements.txt

# To install the minimal dependencies for production use (i.e., what is installed with ``pip install herepy``) run:
$ pip install -Ur requirements.txt
```

## Loading Libraries & Data

# Import libraries: 
import time
import requests
import pandas as pd
import datetime as dt
import numpy as np

import herepy
import matplotlib.pyplot as plt

geocoderApi = herepy.GeocoderApi('INSERT API KEY HERE')

closures = pd.read_csv('../datasets/account_tweets.csv')
address = pd.read_csv('../datasets/address.csv')
address.head()

## Coordinates Acquisition

#### HERE Geocoder API  
Is the wrapper for HERE Geocoder API, to use this wrapper and all other wrappers you need an API key which you can get from HERE Developer Portal.

Initiation of GeocoderApi  

```Python
import herepy

geocoderApi = herepy.GeocoderApi('api_key')
```

Geocoding given search text
```Python
response = geocoderApi.free_form('200 S Mathilda Sunnyvale CA')
```
Geocoding given search text with in given boundingbox   
```Python
response = geocoderApi.address_with_boundingbox('200 S Mathilda Sunnyvale CA',
                                                [42.3952,-71.1056],
                                                [42.3312,-71.0228])
```
Geocoding with given address details
```Python
response = geocoderApi.address_with_details(34, 'Barbaros', 'Istanbul', 'Turkey')
```

Geocoding with given street and city   
```Python
response = geocoderApi.street_intersection('Barbaros', 'Istanbul')
```

Define a function to get coordinates from each location:

def geocoder(address):
    response = geocoderApi.free_form(address)
    return response.Response['View'][0]['Result'][0]['Location']['DisplayPosition']

Using `try` `except` to avoid errors:

# Create a new coordinate column:
address['coor'] = ''
for i in list(address.index):
    try: 
        address['coor'][i] = geocoder(address['address'][i])
    except: 
        pass

# It has coordinates as form {'Latitude':xxx, 'Longitude':xxx}:
address.head()

address.shape

# Drop all 'pass' coordinates:
address.drop(address.loc[address['coor']==''].index, inplace = True)
address.shape

# Create two new columns for 'latitude' and 'longitude':
address['latitude'] = ''
address['longitude'] = ''
for i in list(address.index):
    address['latitude'][i] = address['coor'][i]['Latitude']
    address['longitude'][i] = address['coor'][i]['Longitude']

# Keep only 'latitude' and 'longitude' columns
address.drop(columns=['coor'], inplace = True)

# Get date column from previous data frame:
address['date'] = ''
for i in list(address.index):
    address['date'][i] = closures['date'][i]

address.head()

# Export data:
address.to_csv('../datasets/text_coor_date.csv', index=False)

## Source  

- [HERE.com](https://www.here.com/)
- [HerePy GitHub](https://github.com/abdullahselek/HerePy)
